remove connect() in column_settings
