+++ app/javascript/mastodon/features/notifications/components/column_settings.js
@@ -1,44 +1,22 @@
 import React from 'react';
-import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import ImmutablePropTypes from 'react-immutable-proptypes';
-import { defineMessages, injectIntl, FormattedMessage } from 'react-intl';
-import { clearNotifications } from '../../../actions/notifications';
-import { openModal } from '../../../actions/modal';
+import { FormattedMessage } from 'react-intl';
 import ColumnCollapsable from '../../../components/column_collapsable';
 import ClearColumnButton from './clear_column_button';
 import SettingToggle from './setting_toggle';
-
-const messages = defineMessages({
-  settings: { id: 'notifications.settings', defaultMessage: 'Column settings' },
-  clearMessage: { id: 'notifications.clear_confirmation', defaultMessage: 'Are you sure you want to permanently clear all your notifications?' },
-  clearConfirm: { id: 'notifications.clear', defaultMessage: 'Clear notifications' },
-});
 
 class ColumnSettings extends React.PureComponent {
 
   static propTypes = {
     settings: ImmutablePropTypes.map.isRequired,
-    dispatch: PropTypes.func.isRequired,
     onChange: PropTypes.func.isRequired,
     onSave: PropTypes.func.isRequired,
-    intl: PropTypes.shape({
-      formatMessage: PropTypes.func.isRequired,
-    }).isRequired,
+    onClear: PropTypes.func.isRequired,
   };
 
-  handleClear = () => {
-    const { dispatch, intl } = this.props;
-
-    dispatch(openModal('CONFIRM', {
-      message: intl.formatMessage(messages.clearMessage),
-      confirm: intl.formatMessage(messages.clearConfirm),
-      onConfirm: () => dispatch(clearNotifications()),
-    }));
-  }
-
   render () {
-    const { settings, intl, onChange, onSave } = this.props;
+    const { settings, onChange, onSave, onClear } = this.props;
 
     const alertStr = <FormattedMessage id='notifications.column_settings.alert' defaultMessage='Desktop notifications' />;
     const showStr  = <FormattedMessage id='notifications.column_settings.show' defaultMessage='Show in column' />;
@@ -47,7 +25,7 @@ class ColumnSettings extends React.PureComponent {
     return (
       <div>
         <div className='column-settings__row'>
-          <ClearColumnButton onClick={this.handleClear} />
+          <ClearColumnButton onClick={onClear} />
         </div>
 
         <span className='column-settings__section'><FormattedMessage id='notifications.column_settings.follow' defaultMessage='New followers:' /></span>
@@ -87,4 +65,4 @@ class ColumnSettings extends React.PureComponent {
 
 }
 
-export default connect()(injectIntl(ColumnSettings));
+export default ColumnSettings;
