+++ app/javascript/mastodon/locales/hr.json
@@ -107,7 +107,6 @@
   "notifications.column_settings.reblog": "Boosts:",
   "notifications.column_settings.show": "Prikaži u stupcu",
   "notifications.column_settings.sound": "Sviraj zvuk",
-  "notifications.settings": "Postavke rubrike",
   "onboarding.done": "Učinjeno",
   "onboarding.next": "Sljedeća",
   "onboarding.page_five.public_timelines": "The local timeline prikazuje javne postove svih na {domain}. Federalni timeline pokazuje javne postove svih sa {domain} domena koje slijediš. To je sjajan način da otkriješ nove ljude.",
