+++ app/javascript/mastodon/locales/no.json
@@ -107,7 +107,6 @@
   "notifications.column_settings.reblog": "Fremhevet:",
   "notifications.column_settings.show": "Vis i kolonne",
   "notifications.column_settings.sound": "Spill lyd",
-  "notifications.settings": "Kolonneinstillinger",
   "onboarding.done": "Ferdig",
   "onboarding.next": "Neste",
   "onboarding.page_five.public_timelines": "Den lokale tidslinjen viser offentlige poster fra alle på {domain}. Felles tidslinje viser offentlige poster fra alle som brukere på {domain} følger. Dette er de offentlige tidslinjene, et fint sted å oppdage nye brukere.",
