+++ app/javascript/mastodon/locales/uk.json
@@ -107,7 +107,6 @@
   "notifications.column_settings.reblog": "Передмухи:",
   "notifications.column_settings.show": "Показати в колонці",
   "notifications.column_settings.sound": "Відтворювати звук",
-  "notifications.settings": "Налаштування колонки",
   "onboarding.done": "Готово",
   "onboarding.next": "Далі",
   "onboarding.page_five.public_timelines": "Локальна стрічка показує публічні пости усіх користувачів {domain}. Глобальна стрічка показує публічні пости усіх людей, на яких підписані користувачі {domain}. Це публичні стрічки, відмінний спосіб знайти нових людей.",
